Soften dependency of linotpd package on libapache2-mod-wsgi.
On Debian Jessie (at least), the libapache2-mod-wsgi and
libapache2-mod-wsgi-py3 packages conflict and cannot be installed at
the same time. This creates a problem when on the same system there is
Python 3 code that is to be run under mod_wsgi. With linotpd's hard
dependency on libapache2-mod-wsgi, the package conflict causes the
linotpd package to be uninstallable on such systems, which sucks even
if one doesn't in fact want to use libapache2-mod-wsgi to run linotpd
(one might instead use gunicorn or uwsgi).

To mitigate this problem, we let the linotpd package depend on the
virtual “httpd-wsgi” package as well as libapache2-mod-wsgi. This
means that if no WSGI-capable HTTP server is installed (or about to be
installed) when the linotpd package is about to be installed,
libapache2-mod-wsgi will also be installed, but if a WSGI-capable
server such as gunicorn already exists on the system (or will be
installed at the same time), then that package will be sufficient to
fulfil the dependency.

Change-Id: Ifdd22b7c244562ad329cdf33d22bcae402162851