Use HTTP_AUTHORIZATION to determine login name for Basic auth.
When running LinOTP behind an Apache reverse proxy, the REMOTE_USER
environment variable is not set. In order to let LinOTP figure out
who the authenticated user is in this case, we use the
HTTP_AUTHORIZATION header even when Basic authentication is used from
LinOTP's point of view. (With form-based authentication in Apache,
this header is available if we specify “AuthFormFakeBasicAuth On”.)
Previously, HTTP_AUTHORIZATION was only used to deal with Digest
authentication.

Change-Id: Id9266c85f88c0c5171f75aeff6f5b3cfb72c99ab