Prefer HTTP_X_FORWARDED_HOST to HTTP_HOST for logout_url.
When LinOTP is run behind a reverse proxy, the HTTP_HOST value refers
to the actual URL for LinOTP rather than the externally visible one.
This causes the logout URL (as used by the “Log out” button on the
management web site) to be wrong. We mitigate this problem by checking
for the HTTP_X_FORWARDED_HOST header – which contains the value of
the HTTP “Host:” header as given by the client and seen by the reverse
proxy – in preference to HTTP_HOST if it exists.

Change-Id: I411929fee2d804ad91f0f7c5586becb65bb3e643