[LinOTP] Manage-UI: Modifications to ease testing
The name of a resolver was moved to dedicated span so that you can select
it with ">.name". This was necessary because the _managed_ tag was added.
This simplifies selecting different parts with jquery as well as with the
selenium integration tests.

The server request indecater removal was moved to the end of the callback
function so that it is a save indicator for the selenium tests to wait
for the results.

Minor style changes to move the managed tag to the right of the resolver
list.

Change-Id: Iaf9185e12833cf5dd293896e580c78ffeae3e697