[LinOTP] support for managed providers
managed provider are restricive so that they cant be deleted
or modified easily - it is required to provide a password as
manage parameter. The provider could only be modified or deleted
if the provided password is the same as the initial one.

As we have a long standing nameing conflict with the lib crypt
which could not be resolved, the linotp crypt.py was renamed
to crypto and instantiated as a dedicated module in preparation
for the refactoring of the crypto api

Change-Id: Ia3744ee0212ec8347c48c1337d44e617454396d1