[LinOTP] Manage-UI: Refactored dialog button icon generation
The generic method to add icons dependent to the buttons
context was upgraded to an jquery function and is now called
on the dialogs selector instead of the global document.

This way, we save up to 75 % dialog loading time in the web ui.

Additionally some not covered buttons have proper icons now and
in the case of a translated ui, it is ensured, that the icons
are generated before translation to not lose the context.

Change-Id: I63da4c40cdb457e05cf7d48699847e7758a5c5f5