Hardcode Apache LinOTP ports for Docker containers.
When running on Docker, it is easier to use Docker's port remapping
facility than to reconfigure the actual image. Therefore it makes
sense to fix the ports in the image such that port 80 is used for
unencrypted access to LinOTP, port 443 is used for TLS-based access,
and port 81 is used for health checks. There is a separate facility
to enable an HSTS header if a preceding load balancer accepts HTTPS
requests and forwards them to the container as HTTP (in which case we
may still want to enable HSTS).

Apache listens on ports 80 and 443 by default; we add a “Listen”
directive for port 81 in the Apache LinOTP configuration file for
simplicity (we would otherwise need to add another template for
Debian's ports.conf file).

Refers to LINOTP-343 .

Change-Id: I4baa90f7c2d2dd879d37b7fede6678bb1b49a2b7