[LinOTP] patch the search for user in tokenlist
the token search for a user supports wilcdard search. The detection
if its a wildcard search is determined by looking for '*' as multiple
character search pattern and '.' for single wildcard character, though
the single char wildcard is not evaluated.
As the wildcard serach is very expensiv the '.' for wildcard search
detection is now removed.

refers LinOTP-175

Change-Id: I0028318107c0007e4c9b16e341172f59a99a9004