[LinOTP] Manage-UI: Fix Buttons not converted to themed buttons in tabs
Buttons defined in async loaded manage tabs were not converted to themed
buttons with icons. The buttons that had icons assigned now load and new
buttons are themed with icons.

Therefor a new jquery function was added that possibly can convert more
components in the future. It is called on an jquery selector to find
and convert components in it.

Change-Id: Ibedd1c4a9b21efc6b3cb227e802d8c4d145b9ea5