[LinOTP] Fix for to strong user / owner compare
in the scope of selfservice before any operation on a token, the
token ownership is compared with the user. For historical reasons
the identification of the resolver has 3 entries: the resolver class
appended with the config identifier, the user id and an additional
entry to displaying a resolver specific hint. During the migration
(tools->migrate resolver) the resolver specific hint is fed with an
different value than the native one.
With this patch, the comparrison of the 3rd, not required field is
now omited.

refers LinOTP-493

Change-Id: I09b702d419b80e919cc7c0d81a01955e11fba13d