[LinOTP] Fix for missing hsm parameter in pylons middleware
In selfservice the repoz.who middleware is used for authentication
which has no pylons request context c. In some cases c does
not contain any hsm object which is needed to decrypt and encrypt
password and pin.

Overall with this patch it is intented to get rid of the pylons
request context c and rely only on the linotp request context.
To achive this the request context requires as soon as possible
to refer to the hsm, which is done in the base.py.

In addition the instatiation of the server key required for the
qrtoken and the pushtoken has been moved after the request
context initailisation.

refers LINOTP-531

Change-Id: I0a17248b6143c316275965b1d44cf31027142a00