[LinOTP] Fix successfull authorization for same check_t requests
If a challenge is triggerd and the respond to
check_t will be send twice at the same time, LinOTP will
return two times a successful (positive) answer to the
clients. This fix will lock the 'get challenge for
transaction id' until it is ended. So the second request
will not get the challenge until the first request is finshed.

refers LINOTP-532

Change-Id: I8a0939a26c96c7c0878317bcec69860b61d509e2