Allow setting log levels via environment variables in Docker.
This change lets you say

  $ make docker-run-linotp-sqlite LINOTP_LOGLEVEL=DEBUG

in order to let the general LinOTP log level default to “DEBUG”.

In fact we enable three new environment variables in the LinOTP
configuration file template that dockerfy reads:

  LINOTP_LOGLEVEL controls general LinOTP logging output.
  LINOTP_CONSOLE_LOGLEVEL controls what parts of the general LinOTP
    logging output (as enabled by LINOTP_LOGLEVEL) will end up on
    LinOTP's stdout. (The full firehose will always go to Logstash.)
  SQLALCHEMY_LOGLEVEL controls logging by SQLAlchemy (DEBUG means
    SQL queries and results, INFO means just SQL queries, and WARN
    means neither.)

The (pre-existing) APACHE_LOGLEVEL variable controls the amount of
output Apache logs to its error log. In addition to Apache's own
logging output, this acts as a further filter on material that appears
on LinOTP's console output (and is hence filtered already by
LINOTP_LOGLEVEL and LINOTP_CONSOLE_LOGLEVEL).

It would make reasonable sense to arrange for these environment
variables to work when LinOTP is run inside Rancher (by adding them to
the docker-compose.yml file), but this is outside the scope of this
changeset. Maybe later.

Change-Id: I00947c2b3491251b750350af7973b2e6bae291fd