VoiceToken: Add method getOTP; get_otp
The method getOTP is inherited from hotptoken, which inherit it
from TokenClass. Initially this method calculated the
next otp. For this the hotp serial and count where
fetched from database object. The voice token does not
have a counter. So calculating a next otp is senseless and would
lead to missunderstanding the token. Therefore the
parent getOTP is overritten and raise a NotImplemented
exception.

But we need still a method to get a otp for a random number,
named counter in hotp slang. This will be done with
get_otp()

Change-Id: I235b51f871178df4c2bdd2441f834a29df42f0e9
refers: LINOTP-537