[LinOTP] fix for /validate/check with user and serial parameter
in the case of enrolling and activating an qr ot push token, the
serial number is provided when calling the /validate/check.
In case of a serial number and the user are in the calling args
the evaluation should only be restricted to the token with the
serial number while verifiying the user password/pin.

refers LinOTP-534

Change-Id: I15662787de334cd4c76c34a1963e741b6aa5d689