Debian postinst: simplify code that determines what to do
The previous postinst had a fairly convoluted way of figuring out
which case the postinst was supposed to handle (installation, upgrade,
downgrade, …). We simplify this here to reduce the number of
variables. It turns out the result is only needed in one place right
now, but future improvements might change that.

See LINOTP-499.

Change-Id: I3e342f6cb981b35efc86df5f0c12d8c017a78af7