In linotp, move installation of alembic.ini to setup.py.
The Debian packaging used dh_install to move alembic.ini to
etc/linotp2, when it really should have been moved into place by
setup.py. This didn't work because if setup.py doesn't touch
alembic.ini in the first place, it won't be there for dh_install to
move. This changeset adds alembic.ini to data_files in setup.py, and
that should make the linotp.install (dh_install's configuration file)
unnecessary.

NOTE: Right now there seems to be a bug somewhere that causes
alembic.ini not to be included in the .tar.gz file that "python
setup.py sdist" creates, and that in turn causes an error when "python
setup.py install_data" tries to move alembic.ini into place. Since
alembic support is not required for the upcoming release, we disable
the alembic.ini line in data_files for the time being.

See LINOTP-600.

Change-Id: Ibe22cf4ba0207c1ea3165bec5707331c860a5fcf