[LinOTP] fix: provide request context during response iteration
The userlist request returns a result iterator which is used
eg. for streaming output / HTTP chunked response. When the
response is iterated the request and the request context is
already terminated - it is executed in the pylons context,
where no request context is available anymore.

Part of the request context is the hsm which is used in the
ldap resolver to decrypt the ldap bind password to establish
the ldap connection.

This patch therefore establishes a request_context within the
response iteration so the ldap resolver could connect the ldap
server and retrieve the list of users.

refers LinOTP-587

Change-Id: Ie76543d88f86786fe801ab3958c00e5baddab3f7