[LinOTP] Makefile: changed the name of the `docker-build-selenium` output image
Previously when running integration tests the `selenium_tester`
image would be build and then the very same image would be build as
`integration_selenium_tester`. This happens because `docker-compose`
is prefixing all images it builds with `directory_` in this case
`integration_`.

With this commit the image is only built once, with the proper name.