[LinOTP] tests/integration/Dockerfile: Install missing (`python-`)`packaging` dependency
`packaging` is required to parse the LinOTP version within the tests
but was never installed.  Since `python-packaging` is only available
from stretch onwards the dependency is installed via `pip`.  This is at
best a hack and should probably be fixed when there is an intend to ship
packaged software that can execute the bundled tests.