Fix error where MySQL password gets overwritten during upgrade.
On Debian Jessie, upgrades to LinOTP 2.10 fail because the database
password is erroneously overwritten with the default ("1234") in the
linotp.ini file during the package upgrade. This is due to a typo in
the conditional that decides whether the SQL setup in linotp.ini
should be left alone.

Fixes LINOTP-614.

Change-Id: Ie16bfd9cc984b4b275f9dc5f5864763fa0a426d7