Remove LinOTP dependency on python-repoze.who.
According to Kay Winkler this is no longer required, which is just as
well because the version in Debian Stretch is way newer than 1.9, and
including this dependency makes LinOTP basically uninstallable unless
you find a python-repoze.who package from oldstable, install that
first and put it on HOLD, which is a lot of hassle for a package that
isn't even being used.

See LINOTP-497.

Change-Id: I320a285d8a2c8e7921c953060dfa7f3c39bd493f