"dpkg --compare-versions" safety patch.
In the unlikely case that LinOTP is being configured but Apache2 isn't
installed (which can basically only happen if you install LinOTP using
"dpkg --install --force-depends" but there you are), the LinOTP
postinst can't locate a version number for Apache. Instead of
outputting a syntax error on the "dpkg --compare-versions" call and
then bouncing off into happy chaos, we output a sensible error message
and then bail out.

While we're at it, we're adding safety quotes to a few other "dpkg
--compare-versions" calls, just because we can.

See LINOTP-497.

Change-Id: Iee540707800d4ac3bc95a7960232a40189befe24