Turn linotp-{smsprovider,useridresolver} into "virtual" packages.
In LinOTP 2.10, the previously-separate linotp-smsprovider and
linotp-useridresolver packages have been integrated into the linotp
package, and the linotp package now “Breaks:” and “Replaces:” earlier
versions of these packages. This, however, occasionally creates
problems with other packages that expect to find these as separate
packages that can serve as dependency targets.

We solve this by adding a “Provides:” declaration for these packages
to the linotp package. This turns them into “virtual” packages that
can serve as dependencies for other packages even though they do not
actually exist in their own right. In due course, when the rest of the
ecosystem has been fixed, we may be able to remove this declaration
again.

See LINOTP-625.

Change-Id: I9ae47cacadddf7247947fd82285defa806e1b0b4