Remove calls to systemctl from {pre,post}rm scripts.
These scripts are supposed to work on systems that don't have systemd
installed. It turns out that invoke-rc.d works like it is supposed to;
we just replace the error message by one of our own in case something
goes wrong (because the built-in ones tend to be misleading).

See LINOTP-631.

Change-Id: I7bf36cf49cbf21aef4ccace8db2bf503c291eb3c