[LinOTP] fix for pushtoken unit test
This test failed for libsodiom >=1.0.14.
In old libsodium versions, public key was not checked for validity in
crypto_sign_ed25519_pk_to_curve25519(). To fix this test for newer libsodium
versions, we use a public key that was generated from a secret key and
and hence is valid.

refers LINOTP-642

Change-Id: I9050fba77a6af7e8069fcd8bea5eaef706c93f29