Fix customisation files location for CSS and images
Previously, the only way to apply customisations to LinOTP as packaged
was to copy files directly within the packaging tree itself. This
was under /usr/share/pyshared/linotp for versions 2.9, but changed
for 2.10 to /usr/lib/python2.7/dist-packages/linotp. In this version,
the packaging is changed to separate packaging from customisation files.

The following changes are made:

* All customisations should now be placed under /etc/linotp2/custom-style
* The contents of this directory are made available under the URL
  https://<hostname>/custom
* Previous custom-style.css files will be automatically migrated on
  package upgrade:
    selfservice/custom-style.css -> selfservice-style.css
    manage/custom-style.css -> manage-style.css
    openid/custom-style.css -> openid-style.css
* Users who have placed any additional files on the filesystem, for
  example additional logo or background files, should move these to
  a location under the new path and update references to the file.

LINOTP-729

Change-Id: I7aa20515dc5f7cea1ab323eee76cbfa9a90b1db3