fix: search token list with user principal names
this patch fixes two problems - first a search expression
for username is not split in any case after the '@' sign, only
when the last part might be an existing realm. Second the
evaluation of the regular expression especially with wildcards
now returns only valid matches.

To support better testability the regular expression comparison
has been refactored and isolated in a dedicated function, which
is covered by a unit test.

refers LINOTP-737

Change-Id: I9de40459a7245b22257878731001e1d2aa132257