Add RemoteServiceList and RemoteService for transparent service failover
RemoteServiceList provides a list of RemoteService's that will
transparently continue with another RemoteService whenever one fails.

After an exceeding amount of failures with one service it will be marked
as unavailable for a period of time. Only after that period of time has
passed a service will be retried and eventually marked as available
again.

Refers to LINOTP-754

Change-Id: I567eec097df07d1f6d84efb7938d70ca12ea4711