feature: support check_status wo user parameter
especially for the web ui or with the native api usage
it is very helpfull to be able to query the challenge status
of a token just by the token serial and the token pin.

This patch therefore removes the need of a user. If the user
parameter is given, the pinpolicy of the user is enforced. If no
user is provided the password parameter must satisfy the token pin.

refers to LINOTP-669

Change-Id: I6ecf2e2e779f98ceda80a1736caac1199e4426ac