fix: use a generated public + private key pair in push token test
in the push token unit test a random and not length matching value
was used for the push token uri verification. With libsodium and
pysodium 0.7 though the length and the matching public with
private key is verified.
with the patch in the test we now generate the pub priv key pair
so that the sanity checks in pysodium and libsodium wont fail.

closes LINOTP-782

Change-Id: Ifd87ab9ee0171219e62198dac70d026cfa571cd7