feature: resource scheduler with support for circuit breaking
the resouce scheduler manages a list of resource identifiers -
uris-, which could be maked as blocked for example on a connection
timeout. The resouce will then not be offered for requests until
the blocking delay has expired. In addition a retry could be defined
so that the resource will be connected n-times before it will be
marked as blocked.

this patch only includes the basic class, the resource scheduler
with its unit tests.

refers to LinOTP-648

Change-Id: I590d02588775a618ff5608dd6cf72e4159fa6a6f