fix: user attribute being accessed without check
The `user` variable can be of various types, such as the proper
user object or a falsy value (e.g. None or the empty string).

While most parts of the code would check if the value is truthy, there
was one change in a recent release that wasn't checking the value before
accessing its `realm` property.

The intention of the statement `user.realm or realms or ""` was clearly
to catch issues with the first and second statement, which would
eventually fall back to the empty string. The access of `realm` in the
first term would raise an AttributeError instead of returning a falsy
value and thus causing any of the other terms to be picked.

extracted the abilty to dertmin the login and realm from the options
dict or token context with an unit test

LINOTP-811

Change-Id: I21013f2c1ada3a602e57fb973b5beade21ca4b0a