fix: user compare with ad user
the ad users login name is treated case insensitiv. thus
a comparison between the login name of the user and eg
the token owner is not sufficient. Thus, if the users
exist their userid is used for the comparison.

refers to LINOTP-843

Change-Id: I0384ffe38b933b156f8b698a6d476a64ac1b9e46
(cherry picked from commit df98f61525206ffa52feadeebba6b47edd22835d)