fix: mfa rollout token selfservice login with not mfa_3_fields
in case of mfa and not mfa_3_fields, the login step after the
entered username + password dialog is the token list based
selection of which token should be used for the login. In
this list the rollout token should be provided, while in the
selfservice ui the rollout token should not be shown.

refers to LinOTP-865

Change-Id: I7822c44622e9eff22e1ec242d95e7281c0692844