fix: for wrong validation warning in yubkey enrollment
in form for the yubico enrollment the token id input
was restricted as min=12, which is interpreted as
min integer value. the patch adjusts this to minlength
which sadly is not validated by the html forms native.
the jquery validation for form entries in the enrollement
dialog is not active, so we leave this fix for a later
version.

refers to LINOTP-881

Change-Id: Ie7e8bdd10020ed795207e8985734023972e69655