fix: adjust unit test to catch ProviderNotAvailable
the failover handling requires to work with a generic
exception which explains that the provider is not available
ProviderNoAvailable. This type of exception covers all
type of exceptions, which occure if a Provider is timely
limited not accesible.

The test has been adjusted to check for only the exception
as the message of the lower level exception is platform
dependend.

Change-Id: I35c93eb76fd07aa4bc8194796428677c6f99f96b