fix: for provider test
with using the resource scheduler, in case of a failure test
the provider is blocked for a short delay. Thus the tests had to
be adjusted. The main reason for the failing is that most test
are using the default provider name, which is used if the provider
is installed by using setConfig instead of the setProvider
interface. The patch adjust the tests so that a named default
provider is set for every test.

Change-Id: Id47fbae16d34de1af8cf5cbe27501dcddd710396