feat: understandable error message if token is already locked
in case the of a validate/check the token is read for update.
in case of a paalell request the sql exception ResourceClose
is thrown and results in a non understandable audit entry.
this patch catches the exception and extends the description
so that it describes the error condition more understandable.