chore: totp token verification
the totp time and time shift is verified in this patch. for
documentation of totp time handling algorithm a unit test
and a functional test is deployed

remark: the time2counter and counter2time has been adjusted
to start precisley with the given timestamp or counter instead
of one off - with no impact of the correctnes as linotp always
verifies ranges

remark: for better understanding of the timeshift, the time
delta was reverse mapped to the time slot slices.