chore: pin the docker-compose version to 1.22.0
While this doesn't give us a cryptographically secured version pinning
with an exact hash this should get us pretty close to a reliable way of
installing (almost) the same docker-compose on each pipeline run.

We might want to stick with whatever the Debian repository for some
version offers but that is something for another commit.