feature: Add `linotp.flap` module, to remove Pylons references.
The `linotp.flap` module imports everything we need from `pylons` and
`webob`. Other parts of LinOTP can get their Pylons stuff from this
module instead of importing from `pylons.*` directly.

The advantage of this is that we can then tweak the `linotp.flap`
module to introduce Flask stuff and the rest of the code will pick it
up. For example, right now users of `linotp.flap` will pick up the
Pylons `request` object, but we could have a lookalike that is based
on the Flask `request` object, in order to move from Pylons to Flask
incrementally. Or we could simply consider `from linotp.flap import …`
as a marker that something needs to be ported here.

Note that Pylons and Flask are different enough that, e.g., the
`controllers/` files will probably need major surgery, unless somebody
wants to implement a Pylons base controller that is really based on
Flask.