refactor: Rewrite LinOTP to use `linotp.flap` instead of Pylons.
This huge changeset removes all references to Pylons and replaces them
with references to `linotp.flap`. The only exceptions are some
low-level routines in `websetup.py`, which we will need to rewrite in
any case, and some tests ditto. We're still passing the test suite.

At this point we could theoretically rewrite `linotp.flap` to replace
all its Pylons `import`s with code based on Flask and be done. It's
probably easier to rewrite the actual code, such as the controllers,
to deal with Flask directly.