feature: Add Flask-style initialisation and basic configuration.
This changeset adds a `linotpd/src/linotp/app.py` module which
contains a Flask-style application factory function, and a
`linotpapp.py` file which contains a basic application driver. The
Flask development server can then be started using

    $ FLASK_APP=linotpapp flask run

The `linotpd/src/linotp/settings.py` file contains a basic set of
configuration "environments" which can be addressed by defining the
`FLASK_ENV` environment variable. These include `development`,
`testing`, and `production`. If `FLASK_ENV` is not set, the default is
`default`, which is identical to `development`.

Note that it is customary to include the application factory function
in the `__init__.py` file, but we don't want to do this because this
would interfere with the package building process, where LinOTP's
`__init__.py` is imported from `setup.py`. It would also be more
obvious to call the application driver file `linotp.py`, but that
would lead to problems because we also have the `linotp` Python
package directory. Hence `app.py` and `linotpapp.py`. Sue us.