feature: Introduce `/healthcheck` API endpoint.
This adds an API endpoint to LinOTP which will return a JSON
dictionary of the form

  {'status': 'alive', 'version': '2.10.7', 'uptime': 1.234}

where `version` is the LinOTP version number and `uptime`
the (fractional) number of seconds since this instance of LinOTP was
started.

We're using this to run basic tests on the new Flask-based
infrastructure, but it could come in useful later, e.g., to support
health checks in a Docker infrastructure.