test: Add basic infrastructure for Flask tests.
The current LinOTP test suite is married to Pylons, which makes
testing Flask functionality difficult. Until we can excise all the
Pylons stuff, we use a separate set of tests (based on pytest and
pytest-flask) to test Flask-specific parts of the
infrastructure. Eventually these tests will be made part of the main
test suite.