feature: Add support for "controllers" (class-based blueprints).
This changeset adds tentative support for "controllers", which are
essentially class-based blueprints with automated dispatch to view
functions which are methods. Compared to the standard Flask approach
of using top-level view functions, this lets us use inheritance so
controllers can share code of common interest. Controller classes are
derived from the `BaseController` class.

When a "controller" module is loaded (e.g, `foo`), we look for a class
called `FooController` and register routes to all its methods whose
names do not start with an underscore, with a default URL prefix
derived from the module name (here, `/foo`). The default class name
prefix and default URL prefix may be overridden in the list of
controllers to be imported (which is a configuration setting).

View methods are assumed to accept the `GET` HTTP method unless a
different set of HTTP methods is specified using the `@methods`
decorator, as in

    class FooController(BaseController):
        ...
	@methods(['GET', 'POST'])
	def bar(self):
	    ...