refactor: Moved Flask `BaseController` to `controllers/base.py`.
In order to slim down `controllers/__init__.py`, the `BaseController`
class in that file is folded into the
`controllers.base.BaseController` class. We also move the
`ControllerMetaClass` and the `methods` decorator over to
`controllers/base.py`. This means that `controllers/__init__.py` is
now essentially empty.

Note that we change the metaclass to do its magic only for classes
that are not called `BaseController`. This is to avoid having to
rename all the utility methods in the `BaseController` class just so
we don't generate API routes to them.