refactor: Move database initialisation to Flask app factory.
This used to be done in the `linotp.controllers.base.setup_app()`
function but is now part of the Flask `linotp.app.create_app()`
factory function (or, more precisely, a function called from that
function). At some point we will want to rewrite this to use
Flask-SQLAlchemy and Flask-Migrate.

Note that we're no longer initialising the database connection in
`linotp.config.environment.load_environment()`, which ended up as a
per-request thing in the Flask port. We can get away with this because
the database URL comes from the Flask-side configuration and we don't
need to take into account the various oddities of Pylons application
startup.