refactor: Move database default setting out of `controllers/base.py`.
This code has no business being in `controllers/base.py` since it must
be executed only once when a LinOTP installation is new (or being
upgraded). We move the `set_defaults()` and `set_config()` functions
to a new `linotp/config/defaults.py` file and the code to invoke these
to the Flask application factory function.