feature: Install additional route for endpoints with `id` parameter.
Some endpoint methods support both `/foo/bar` and `/foo/bar/ID`
URLs (where `ID` refers to an object of some description). This used
to be implemented in Pylons in a way that is not suitable to describe
in polite company. With Flask, we introspect the method signatures and
if we detect a method that has two parameters, the second of which is
an optional one called `id`, we install an additional `…/ID` route as
above.