port: Serve static content under '/static'
Flask expects all static files to be available under one toplevel
URL. This makes sense from a DevOps perspective, because it makes it
easy to configure a webserver such as Apache to serve all content
under one path. Previously this was scattered in a few places
(css, images, js, etc.). So given it is harder to administer a system
containing multiple static paths, we move everything under a single
path, '/static'. In order to avoid having to move files around in git
at this stage, we serve the static files from the same location as
before, which is the folder 'public'.