refactor: Move loading the global object to the app factory.
Again this is something that we really want to do once per application
startup rather than once per request. We're still doing it once per
request because it must hang off the `flask.g` variable for now, but
at least we have it in the application factory where it belongs.