fix: loading the token templates adjusted to use an relative path
in the flask port there is currently only one template directory
defined. but as some tokens are sole packages e.g the pushtoken
the mako files are not found via the getClassInfo defintion.
thus the defintion in the getClassInfo has been adjusted with
an relative path for the mako file.