test: Enable (functional) tests for validate and admin controllers.
We want to see how we fare in the functional tests, so we stop
blacklisting the `validate` controller. This controller uses
operations from the `admin` controller in various places, so we need
to unblock that controller too, otherwise the `xfail`-ing `admin`
operations would cause the validate operations to `xfail`, too.

For now we don't really care what happens to the `admin` tests
themselves.

We also tweak the `TestValidateController.test_unicode()` test to make
it succeed. Right now the only test that still fails is
`TestValidateController.test_auth_info_detail_radiotoken()`, and
somebody else will need to look into that one.