test: Unify app setup for pytest fixtures and TestController
* Removed the need for TestController to override __init__

By using the pytest app fixture, we can remove some setup code
from the old TestController and make use of the same features
in the fixture setup. Now we can start each functional test with
a completely clean database, which means tests cannot fail due to
an issue with an unrelated test.

Related to LINOTP-1086