fix: Day-password token expects `secretObject` key to be `bytes`.
This was originally mocked to return a `str` but it must be `bytes`
for `zerome()` to work, so we proceed under the assumption that that
was what was intended.

Note that we must encode the `date_string` parameter to `bytes` so we
can concatenate it to `self.key`. While we're futzing around with this
we remove the dependency on the `binascii` module.