replace jquery 1.12 with 3.4 and jquery-migrate
JQuery 1.12 is no longer maintained and has to be replaced with a
newer version. This commit adds the jquery-migrate library that
restores deprecated APIs. This way we can support older libraries with
a newer jquery version.

The migration library also aids the developer by throwing warnings if
deprecated APIs get used. To enable this output, replace the minified
jquery-migrate with the non-minified version. Once all deprecated calls
are removed, the migration library can be removed.

This closes LINOTP-1173