fix: for qrtoken with content-type requesting signed response
when content type is pairing the protocol specifies that
the server must send a hmac based signature with the
response. in this case the signature is already binary and not
an empty string ''. to handle both cases the same way the
initialisation is done by an empty byte sting b''.