test: Fix spurious `assert` statements in `test_selfservice.py`.
Various `assert` statements tried to check that a string containing a
JSON reply from the server contained a nested dictionary member with a
specific string value. Unfortunately the value tested for had a
trailing comma, which is only included if the member in question is
not the last member of the dictionary, but it was. Bummer.

A few other `assert` statements checked for strings with internal
quoting which needed to be adjusted.