fix: for get and delete in config class
the has_key method was replaced in python3 by the in operator.
the in operator of the config class is tolerant wrt if the
'linotp.' prefix is given or not.
but in case of a delete or get the exact key to return or
delete is required to know.
Thus we have to do a positiv existance test for prefixed key
first thus beeing sure that the right key is used for the
delete and get in the parent dict implementation.