fix: Avoid extraneous UTF-8 encoding in `ocra2token.py`.
By default, JSON encoding will fail for `bytes` objects. We must
therefore ensure that stuff which will go to JSON is not otherwise
already encoded to `bytes` before the JSON encoder gets to have a go
at it. This applies to the shared secret, which is of type `bytes`
already and must be converted to `str`, and the callback URL which is
a `str` and can stay that way.