chore: clarify the SecretObject methods api
the SecretObject is used to provide the api for the end user
to implement encryption / decryption of pin and passwords or
hashing of pins. This level of api should work mostly on str
level.
In addition to clarify the apis an additional api was introduced
for the crypted or hashed pin comparison, which takes the output
of a former encryptPin result as input. This api extension will
allow to support the position independend pin comparison to be
localted in the crypto lib.